requirements.txt
azure-core==1.18.0
azure-storage-blob==12.9.0
pandas==1.3.3
plotly==5.3.1
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.3.0
tenacity==8.0.1